camerabin: update framerate choosing logic
When there is no proper caps which have the desired aspect ratio,
gst_camerabin_set_allowed_framerate returns caps without framerate. This may
cause a problem in certain cases. Because the capsfilter uses last used
framerate if there is no framerate specified in the new caps. So the framerate
choosing logic is updated to choose the framerate as long as there is
an intersect between the allowed and the desired resolution.